+++ tools/odrive/enums.py
@@ -18,6 +18,7 @@ GPIO_MODE_PWM0                           = 10
 GPIO_MODE_ENC0                           = 11
 GPIO_MODE_ENC1                           = 12
 GPIO_MODE_ENC2                           = 13
+GPIO_MODE_MECH_BRAKE                     = 14
 
 # ODrive.Can.Protocol
 PROTOCOL_SIMPLE                          = 0