+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased Features
 Please add a note of your changes below this heading if you make a Pull Request.
+### Added
+* [Mechanical brake support](docs/mechanical-brakes.md)
 
 ### Changed